requirements.txt
asgiref
Django==5.0.4
djangorestframework==3.15.1
psycopg2
# psycopg-binary
sqlparse==0.5.0
typing_extensions==4.11.0
